+++ front/main/app/routes/application.js
@@ -136,9 +136,8 @@ export default Route.extend(
 				}
 
 				if (info.article) {
-					this.set('adsHighImpact.renderInterstitial', true);
-
-					//this.transitionTo('wiki-page', info.article + (info.hash ? info.hash : ''));
+					this.get('adsHighImpact').reload();
+					this.transitionTo('wiki-page', info.article + (info.hash ? info.hash : ''));
 				} else if (info.url) {
 					/**
 					 * If it's a jump link or a link to something in a Wikia domain, treat it like a normal link
