+++ front/main/app/templates/components/ads/invisible-high-impact-2.hbs
@@ -1,7 +1,8 @@
-<div id="InvisibleHighImpactWrapper" class={{unless renderInterstitial "hidden"}}>
+<div id="InvisibleHighImpactWrapper" class="hidden">
 	<div class="background"></div>
 	<a class="close">
 		<div class="close-button"></div>
 	</a>
 	<div class="label">Advertisement</div>
+	<div id="{{name}}" class="ad-slot hidden"></div>
 </div>
